[icons] Move icon utilities to own module
