+++ lib/js/components/animation-fundamentals.mjs
@@ -2,10 +2,6 @@
 /* jshint -W008 */ // leading dot in decimals...
 
 import {
-    createIconAndLabel
-} from '../domTool.mjs';
-
-import {
   //  Path
   //, getEntry
   //, ForeignKey
@@ -35,6 +31,10 @@ import {
   // , UILineOfTextInput
   // , MoveItemsInListButton
 } from './generic.mjs';
+
+import {
+    createIconAndLabel
+} from './icons.mjs';
 
 export * from './animation-t-generator.mjs';
 export * from './ui-time-control-circle.typeroof.jsx';
