+++ lib/js/components/basics.mjs
@@ -1,9 +1,7 @@
 /* jshint esversion: 11, browser: true, unused:true, undef:true, laxcomma: true, laxbreak: true, devel: true */
 /* jshint -W008 */ // leading dot in decimals...
 
-import DOMTool, {
-    createIcon
-} from '../domTool.mjs';
+import DOMTool from '../domTool.mjs';
 
 import {
     Path
@@ -15,6 +13,9 @@ import {
   , FreezableMap
 } from '../metamodel.mjs';
 
+import {
+    createIcon
+} from './icons.mjs';
 
  // To mark the update strategy of the widget
 export const UPDATE_STRATEGY_SIMPLE = Symbol('UPDATE_SIMPLE')
