+++ lib/js/components/generic.mjs
@@ -10,9 +10,7 @@ import {
 } from '../metamodel.mjs';
 
 import {
-    HTMLString,
-    createIconAndLabel,
-    createLabelAndIcon
+    HTMLString
 } from '../domTool.mjs';
 
 import{
@@ -28,6 +26,11 @@ import {
   , UPDATE_STRATEGY_NO_UPDATE
   , UPDATE_STRATEGY
 } from './basics.mjs';
+
+import {
+    createIconAndLabel
+  , createLabelAndIcon
+} from './icons.mjs';
 
 export class GenericSelect extends _BaseComponent {
     // jshint ignore:start
