+++ lib/js/components/layouts/stage-and-actors.mjs
@@ -44,6 +44,10 @@ import {
 } from '../generic.mjs';
 
 import {
+    createIcon
+} from '../icons.mjs';
+
+import {
     _BaseLayoutModel
 } from '../main-model.mjs';
 
@@ -147,10 +151,6 @@ import {
 import {
     UIshowProcessedProperties
 } from '../processed-properties.mjs';
-
-import {
-    createIcon
-} from '../../domTool.mjs';
 
 const StageSizeNumberModel = _AbstractNumberModel.createClass(
         'SizeNumberModel'
