+++ lib/js/components/layouts/videoproof-array-v2.mjs
@@ -16,10 +16,6 @@ import {
 } from '../../util.mjs';
 
 import {
-    createIcon
-} from '../../domTool.mjs';
-
-import {
     _BaseContainerComponent
   , _BaseDynamicCollectionContainerComponent
   , _BaseComponent
@@ -31,6 +27,10 @@ import {
   , UICheckboxOrEmptyInput
   , Collapsible
 } from '../generic.mjs';
+
+import {
+    createIcon
+} from '../icons.mjs';
 
 import {
     _BaseLayoutModel
