+++ lib/js/domTool.mjs
@@ -22,18 +22,6 @@ ValueError.prototype.constructor = ValueError;
 // `span`s for icons loaded from the Material Symbols font.
 export class HTMLString extends String {}
 
-export function createIcon(name) {
-    return new HTMLString(`<span class="material-symbols-outlined">${name}</span>`);
-}
-
-export function createIconAndLabel(iconName, label) {
-    return new HTMLString(`${createIcon(iconName)} ${label}`);
-}
-
-export function createLabelAndIcon(label, iconName) {
-    return new HTMLString(`${label} ${createIcon(iconName)}`);
-}
-
 export default class DOMTool {
   constructor(document){
         this.document = document;
